Modified script to reflect exit code accurately
The final part of the script functionality, in which the downloaded artifact is moved from its temporary staging area to its final deployment location, was masking a failed curl call.  As such, the move command was moving a 0-byte length file to the final deployment location and the script was exiting with a '0' status.  Puppet therefore erroneously thinks all is well.  With this fix the script with now bubbles up the return status of a failed curl call and prevents the empty file from being moved.